BookTransaction

+++ src/main/resources/db/migration/V2_2__issued_book.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS issued_books(
   id bigint(20) NOT NULL AUTO_INCREMENT,
   bookId bigint(20) NOT NULL,
-  employeeId INTEGER NOT NULL,
+  employeeId bigint(20) NOT NULL,
   issueDate DATE,
   returnedDate DATE,
   isActive BOOLEAN,
